# front_start/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для управления квантовыми задачами и гейтами",
        "title": "Quantum Tasks API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/draft/gates/{id}": {
            "post": {
                "description": "Добавляет указанный гейт в текущую черновую задачу пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gates"
                ],
                "summary": "Добавить гейт в черновик задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гейта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_TaskServiceLink"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID гейта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при добавлении гейта в задачу",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/gates": {
            "get": {
                "description": "Возвращает список всех гейтов. Поддерживает фильтрацию по названию.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gates"
                ],
                "summary": "Получить список гейтов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию гейта (поиск по подстроке)",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_app_handler.DTO_Resp_Gate"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новый квантовый гейт с указанными параметрами.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gates"
                ],
                "summary": "Создать новый гейт",
                "parameters": [
                    {
                        "description": "Данные нового гейта",
                        "name": "gate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Req_GateCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/front_start_internal_app_ds.Gate"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные запроса",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при создании гейта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/gates/{id}": {
            "get": {
                "description": "Возвращает полную информацию о гейте по его идентификатору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gates"
                ],
                "summary": "Получить гейт по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гейта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/front_start_internal_app_ds.Gate"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Гейт не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет существующий гейт по ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gates"
                ],
                "summary": "Обновить гейт",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гейта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные гейта",
                        "name": "gate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Req_GateCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/front_start_internal_app_ds.Gate"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные запроса или ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при обновлении гейта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет гейт по ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gates"
                ],
                "summary": "Удалить гейт",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гейта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_SimpleID"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при удалении гейта",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/gates/{id}/image": {
            "post": {
                "description": "Загружает изображение гейта и обновляет URL в базе данных.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gates"
                ],
                "summary": "Загрузить изображение для гейта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гейта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение гейта",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_UploadImg"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID или отсутствует файл",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при загрузке изображения",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/quantum_task/current": {
            "get": {
                "description": "Возвращает ID текущей черновой задачи и количество добавленных в неё гейтов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuantumTasks"
                ],
                "summary": "Получить информацию о текущей черновой задаче",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_CurrTaskInfo"
                        }
                    },
                    "500": {
                        "description": "Ошибка при получении данных задачи",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "front_start_internal_app_ds.DegreesToGates": {
            "type": "object",
            "properties": {
                "degrees": {
                    "type": "number",
                    "format": "float32"
                },
                "gate": {
                    "description": "Связи, необходимые для работы сервиса",
                    "allOf": [
                        {
                            "$ref": "#/definitions/front_start_internal_app_ds.Gate"
                        }
                    ]
                },
                "id_gate": {
                    "type": "integer"
                },
                "id_task": {
                    "type": "integer"
                },
                "task": {
                    "$ref": "#/definitions/front_start_internal_app_ds.QuantumTask"
                }
            }
        },
        "front_start_internal_app_ds.Gate": {
            "type": "object",
            "properties": {
                "degrees": {
                    "description": "Несущая связь",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/front_start_internal_app_ds.DegreesToGates"
                    }
                },
                "description": {
                    "type": "string"
                },
                "fullInfo": {
                    "description": "subject area",
                    "type": "string"
                },
                "i0j0": {
                    "type": "integer"
                },
                "i0j1": {
                    "type": "integer"
                },
                "i1j0": {
                    "type": "integer"
                },
                "i1j1": {
                    "type": "integer"
                },
                "id_gate": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "matrix_koeff": {
                    "type": "number"
                },
                "status": {
                    "type": "boolean"
                },
                "theAxis": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "front_start_internal_app_ds.QuantumTask": {
            "type": "object",
            "properties": {
                "conclusionDate": {
                    "type": "string"
                },
                "creationDate": {
                    "type": "string"
                },
                "formedDate": {
                    "type": "string"
                },
                "gatesDegrees": {
                    "description": "Несущая связь",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/front_start_internal_app_ds.DegreesToGates"
                    }
                },
                "id_moderator": {
                    "type": "integer"
                },
                "id_task": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "moderator": {
                    "$ref": "#/definitions/front_start_internal_app_ds.Users"
                },
                "res_koeff_0": {
                    "type": "number"
                },
                "res_koeff_1": {
                    "type": "number"
                },
                "taskDescription": {
                    "description": "Задание пользователя",
                    "type": "string"
                },
                "taskStatus": {
                    "type": "string"
                },
                "user": {
                    "description": "Связь для программы",
                    "allOf": [
                        {
                            "$ref": "#/definitions/front_start_internal_app_ds.Users"
                        }
                    ]
                }
            }
        },
...
        "internal_app_handler.DTO_Req_GateCreate": {
            "description": "Данные для создания нового квантового гейта",
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Quantum Fourier transform gate"
                },
                "full_info": {
                    "type": "string",
                    "example": "Creates superposition states"
                },
                "i0j0": {
                    "type": "integer",
                    "example": 1
                },
                "i0j1": {
                    "type": "integer",
                    "example": 0
                },
                "i1j0": {
                    "type": "integer",
                    "example": 0
                },
                "i1j1": {
                    "type": "integer",
                    "example": 1
                },
                "matrix_koeff": {
                    "type": "number",
                    "example": 0.707
                },
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "the_axis": {
                    "type": "string",
                    "enum": [
                        "X",
                        "Y",
                        "Z",
                        "non"
                    ],
                    "example": "X"
                },
                "title": {
                    "type": "string",
                    "example": "Hadamard Gate"
                }
            }
        },
...
        "internal_app_handler.DTO_Resp_CurrTaskInfo": {
            "description": "Статистика по текущей задаче пользователя",
            "type": "object",
            "properties": {
                "services_count": {
                    "type": "integer",
                    "example": 3
                },
                "task_id": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "internal_app_handler.DTO_Resp_Gate": {
            "description": "Полная информация о квантовом гейте",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Quantum NOT gate"
                },
                "full_info": {
                    "type": "string",
                    "example": "Bit flip gate"
                },
                "i0j0": {
                    "type": "integer",
                    "example": 0
                },
                "i0j1": {
                    "type": "integer",
                    "example": 1
                },
                "i1j0": {
                    "type": "integer",
                    "example": 1
                },
                "i1j1": {
                    "type": "integer",
                    "example": 0
                },
                "id_gate": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "gate_x.png"
                },
                "matrix_koeff": {
                    "type": "number",
                    "example": 1
                },
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "the_axis": {
                    "type": "string",
                    "example": "X"
                },
                "title": {
                    "type": "string",
                    "example": "Pauli-X"
                }
            }
        },
...
        "internal_app_handler.DTO_Resp_SimpleID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_app_handler.DTO_Resp_TaskServiceLink": {
            "description": "Информация о связи между задачей и гейтом",
            "type": "object",
            "properties": {
                "service_id": {
                    "type": "integer",
                    "example": 3
                },
                "task_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "internal_app_handler.DTO_Resp_UploadImg": {
            "description": "Результат загрузки изображения для гейта",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "gate_image.png"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}